requirements.txt
opencv-python==4.8.1.78
ultralytics>=8.3.250
python-telegram-bot==20.7
numpy==1.24.3
Pillow==10.0.1
rdflib==6.3.2
python-dotenv==1.0.0
pydantic==2.5.0
